Upload
The unitTest file is for the region Models part, include the method of show_message and get_major_report_type. This part is completed done.
The Pytest file is for the rest part, I try to use some of the code in test_importFile to do this. However, it does not work well. I think this part is a bit complex for me. I just hand in all I got.